component 분리 , 스타일드 컴포넌트 분리 및 적용

+++ src/pages/Detail.jsx
@@ -5,47 +5,64 @@ import {
   deleteZanNaBiLetter,
   updateZanNaBiLetter,
 } from "store/modules/znbFanLetter";
-import styled from "styled-components";
+import {
+  StLetterCard,
+  StLetterText,
+  StLetterCardOptionButton,
+  StLetterCardSave,
+  StLetterCardCancel,
+  StLetterCardDelete,
+  StLetterCardUpdate,
+} from "../style/LogicalStyle";
 
+// Detail 잔나비 레터 상세페이지 컴포넌트
 const Detail = () => {
   const { id } = useParams();
   const navigate = useNavigate();
-
-  // 수정모드는  첫번째에는 비활성화
   const [isEditing, setIsEditing] = useState(false);
   const [editedContent, setEditedContent] = useState();
-
   const letterData = useSelector((state) => state.zaNaBiLetter);
 
-  // 팬레터 내용을 변경하고 , 삭제한 것을 반영하기 위해서 사용함
   const dispatch = useDispatch();
+
+  const handleSave = () => {
+    const isUpdate = window.confirm("수정하시겠습니까?");
+
+    if (isUpdate === true) {
+      dispatch(updateZanNaBiLetter({ id, editedContent }));
+      setIsEditing(false);
+      navigate("/");
+    } else {
+      setIsEditing(true);
+    }
+  };
+
+  const handleDelete = (id) => {
+    // 삭제 validation 체크
+    const isDelete = window.confirm("삭제하시겠습니까?");
+    if (isDelete === true) {
+      dispatch(deleteZanNaBiLetter(id));
+      navigate("/");
+    } else {
+      return;
+    }
+  };
 
   const handleEdit = () => {
     setIsEditing(true);
-  };
-
-  const handleSave = () => {
-    dispatch(updateZanNaBiLetter({ id, editedContent }));
-    setIsEditing(false);
-    navigate("/");
-
-    if (editedContent === "") {
-      alert("안을 비울수 없습니다.");
-      return;
-    }
   };
 
   const handleCancelEdit = () => {
     setIsEditing(false);
   };
 
-  const handleDelete = (id) => {
-    dispatch(deleteZanNaBiLetter(id));
-    navigate("/");
-  };
-
   return (
     <>
+      {/* 
+      useParam을 이용하여 받은 id값을 통해 필터링 된 새 배열을 
+      map으로 통해 뿌려준것   
+    
+    */}
       {letterData
         .filter((lD) => lD.id === id)
         .map((LD) => {
@@ -63,11 +80,12 @@ const Detail = () => {
                     onChange={(e) => {
                       setEditedContent(e.target.value);
                     }}
+                    required
                   >
                     {LD.content}
                   </StLetterText>
                 ) : (
-                  <p>{LD.content}</p>
+                  <StLetterText>{LD.content}</StLetterText>
                 )}
 
                 <StLetterCardOptionButton>
@@ -106,40 +124,5 @@ const Detail = () => {
     </>
   );
 };
-
-const StLetterCard = styled.p`
-  display: flex;
-  justify-content: space-around;
-`;
-
-const StLetterText = styled.textarea`
-  width: 100%;
-  height: 100px;
-  margin-top: 20px;
-  margin-left: 20px;
-`;
-
-const StLetterCardOptionButton = styled.div`
-  margin-left: 20px;
-  margin-top: 10px;
-`;
-
-const StLetterCardSave = styled.button`
-  margin-right: 10px;
-  border-radius: 5px;
-`;
-
-const StLetterCardCancel = styled.button`
-  border-radius: 5px;
-`;
-
-const StLetterCardDelete = styled.button`
-  margin-right: 10px;
-  border-radius: 5px;
-`;
-
-const StLetterCardUpdate = styled.button`
-  border-radius: 5px;
-`;
 
 export default Detail;
